MainDbContext Complete...

+++ PosLibrary/Model/Entities/Transactions/TransactionPayments.cs
@@ -5,7 +5,7 @@ namespace PosLibrary.Model.Entities.Transactions
     public class TransactionPayments : CommonEntity
     {
         public string ReceiptId { get; set; }
-        public decimal PaymentMethodId { get; set; }
+        public int PaymentMethodId { get; set; }
         public int TransactionHeaderId { get; set; }
         public decimal TotalAmount { get; set; }
         public virtual PaymentMethod PaymentMethod { get; set; }
